Harden latent overflow, bounds, and data-race hazards
- Compute per-cell x per-continuum estimator indices in ptrdiff_t
  instead of int: for very large models the 32-bit product
  nonemptymgi * nbfcontinua_ground can overflow, corrupting the
  photoionisation/bf-heating estimator arrays (rpkt.cc hot path,
  thermalbalance, update_grid, ltepop, grid, radfield, ratecoeff).

- choose_gamma_ray: the cumulative line probabilities reach exactly one
  only up to floating-point rounding, so a random draw could fall past
  the last line and abort. Return the last line instead.

- Restarting: reject gridsave files whose non-thermal excitation list
  exceeds the allocated per-cell capacity (previously an unbounded read
  into shared memory), and write node-shared grid arrays (T_R/T_e/W/T_J,
  thick, nne, nnetot, corrphotoionrenorm) from the node master only,
  matching the guarded pattern in assign_initial_temperatures.

- calculate_decaychain: fail with a clear assertion when two nuclides in
  a chain have equal decay constants (the Bateman form is singular)
  instead of dividing by zero and aborting on a NaN downstream.

- do_packet TYPE_KPKT: assert that the packet is in a non-empty cell
  before indexing thick_allcells, documenting the invariant that the
  equivalent r-packet code checks explicitly.

+++ rpkt.cc
@@ -482,7 +482,7 @@ void update_estimators(const double e_cmf, const double nu_cmf, const double dis
         // because groundcont is sorted by nu_edge ascending, nu_cmf < nu_edge for all remaining items
         return;
       }
-      const int ionestimindex = (nonemptymgi * globals::nbfcontinua_ground) + i;
+      const ptrdiff_t ionestimindex = (static_cast<ptrdiff_t>(nonemptymgi) * globals::nbfcontinua_ground) + i;
 
       if constexpr (USE_LUT_PHOTOION) {
         atomicadd(globals::gammaestimator[ionestimindex],
